move things into its own folders
